Refactored the COSMOS ingestion pipeline and added an app to annotate COSMOS' json files from the command line
